dataSetup - Changed method of reading in images to use PIL instead of torch functions. This is because the test images are 16-bit which torchvision.io.read_image does not accept
loss - L_ssim is clamped to only return values between 0 and 1, although this should have already been the case as that is the range of the ssim function. However I choose to clamp it just in case.

train has been added. I believe there may be something wrong with the loss and ssim, although i do not know yet if it is the functions themselves or the way I choose to call them.